Split XSD from WSDL
