+++ wsdl-model/build.gradle.kts
@@ -3,6 +3,5 @@ plugins {
 }
 
 dependencies {
-    api(libs.serialization.core)
-    api(libs.serialization.xml)
+    api(projects.xsdModel)
 }
