+++ xsd-model/README.md
@@ -1,0 +1,3 @@
+# Module wsdl-model
+
+This module contains the data model to read a wsdl file.
