+++ xsd-model/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    id("compilerModule")
+}
+
+dependencies {
+    api(libs.serialization.core)
+    api(libs.serialization.xml)
+}
